Commented transition component
