+++ components/transition/transition.module.ts
@@ -1,6 +1,7 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
-import {SuiTransition, Transition, TransitionDirection, TransitionController} from './transition';
+import {SuiTransition, Transition, TransitionDirection} from './transition';
+import {TransitionController} from './transition-controller';
 
 @NgModule({
     imports: [CommonModule],
